tests/wpt/web-platform-tests/keyboard-lock/navigator-keyboard-lock.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
'use strict';

promise_test(() => {
  const p = navigator.keyboard.lock(["KeyA", "KeyB"]);
  assert_true(p instanceof Promise);
  return p;
}, '[Keyboard Lock] keyboard.lock');

</script>
